Call

+++ TP3/src/main/java/ar/edu/itba/sia/gae/selectionMethods/Elite.java
@@ -11,7 +11,7 @@ public class Elite extends SelectionMethod {
         Selects the firsts with better fitness.
      */
     @Override
-    public List<Character> select(List<Character> population, int size, int generations) {
+    public List<Character> select(List<Character> population, int size, long generations) {
         List<Character> copy = new LinkedList<>(population);
         copy.sort(Collections.reverseOrder());
         return copy.subList(0, size);
