+++ TP3/src/main/java/ar/edu/itba/sia/gae/selectionMethods/SelectionMethod.java
@@ -3,5 +3,5 @@ package ar.edu.itba.sia.gae.selectionMethods;
 import java.util.List;
 
 public abstract class SelectionMethod {
-    public abstract List<Character> select(List<Character> population, int size, int generations);
+    public abstract List<Character> select(List<Character> population, int size, long generations);
 }
